refactor: move auto prompt caching helpers to utils.py
- Move helper functions out of endpoints.py into
  litellm/proxy/anthropic_endpoints/utils.py
- Use Anthropic types (AllAnthropicMessageValues) instead of
  generic Dict[str, Any]
- Handle content=None case gracefully
- Add test for None content edge case
